# lib/api-spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Valo Pay sandbox API",
    "version": "1.1.0",
    "description": "Valo Pay collections and connected banking sandbox API. All monetary fields are integer minor units (NGN kobo). Real data and all outbound provider instructions are disabled in connected modules."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/v1/actions": {
      "post": {
        "operationId": "performAction",
        "tags": [
          "valopay"
        ],
        "parameters": [
          {
            "name": "merchantId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The lender (a merchant in the API) the request is scoped to; one of the caller's workspace merchants."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Permission or readiness gate blocked"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionInput"
              }
            }
          }
        },
        "summary": "Run a domain action on the lender's state",
        "description": "Every action is audited, most require a reason, and the persona's role applies; the catalogue of actions is in docs/frontend-contract.md."
      }
    },
    "/v1/imports": {
      "post": {
        "operationId": "importRecords",
        "tags": [
          "valopay"
        ],
        "parameters": [
          {
            "name": "merchantId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The lender (a merchant in the API) the request is scoped to; one of the caller's workspace merchants."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Permission or readiness gate blocked"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportInput"
              }
            }
          }
        },
        "summary": "Preview or commit a synthetic CSV import",
        "description": "commit=false validates every row and reports each; commit=true persists all rows or none. syntheticOnly must be true: no real lender data."
      }
    },
...
    "/v1/settings": {
      "get": {
        "operationId": "getSettings",
        "tags": [
          "valopay"
        ],
        "parameters": [
          {
            "name": "merchantId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The lender (a merchant in the API) the request is scoped to; one of the caller's workspace merchants."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Permission or readiness gate blocked"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "summary": "A lender's settings, permissions, integrations, members and calendar",
        "description": "Permissions are those of the caller's current persona."
      },
      "patch": {
        "operationId": "updateSettings",
        "tags": [
          "valopay"
        ],
        "parameters": [
          {
            "name": "merchantId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The lender (a merchant in the API) the request is scoped to; one of the caller's workspace merchants."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Permission or readiness gate blocked"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsInput"
              }
            }
          }
        },
        "summary": "Change a lender's execution settings",
        "description": "Admin only. Send expectedRevision from the settings originally opened; 409 leaves outdated edits unapplied. The revision covers editable preferences and is unaffected by scheduler cursor changes. An identical successful Idempotency-Key replay returns its original result before checking the version."
      }
    },
    "/v1/exports": {
      "post": {
        "operationId": "createExport",
        "tags": [
          "valopay"
        ],
        "parameters": [
          {
            "name": "merchantId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The lender (a merchant in the API) the request is scoped to; one of the caller's workspace merchants."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Permission or readiness gate blocked"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportInput"
              }
            }
          }
        },
        "summary": "Queue a private export",
        "description": "Durably saves a queued export job and returns immediately. Poll its status before downloading. Rendering and private storage run outside the database transaction; retries use the same immutable object key. A record kind, gate pack, billing statement or customer dispute pack supports JSON, CSV or PDF."
      }
    },
...
    "/v1/connected": {
      "get": {
        "operationId": "getConnectedWorkspace",
        "tags": [
          "valopay"
        ],
        "parameters": [
          {
            "name": "merchantId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The lender (a merchant in the API) the request is scoped to; one of the caller's workspace merchants."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectedWorkspace"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Permission or readiness gate blocked"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "summary": "Read the connected workspace",
        "description": "Same-origin, private and tenant scoped. Returns granular consent state, bound sample payment intents, explained credit assessments, independent SME cash planning and closed live gates."
      }
    },
    "/v1/connected/actions": {
      "post": {
        "operationId": "performConnectedAction",
        "tags": [
          "valopay"
        ],
        "parameters": [
          {
            "name": "merchantId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The lender (a merchant in the API) the request is scoped to; one of the caller's workspace merchants."
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200
            },
            "description": "One key per unchanged intention. Retain it after response loss; replay happens before revision checking."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectedActionResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Permission or readiness gate blocked"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectedActionInput"
              }
            }
          }
        },
        "summary": "Perform a synthetic connected-workspace action",
        "description": "Runs inside the existing merchant transaction and audit boundary. Role, purpose, subject, expiry, ownership and version checks apply. Unknown payment outcomes hold retries. No bank, credit bureau, accounting or tax endpoint is called."
      }
    }
  },
  "components": {
    "schemas": {
...
      "RecordData": {
        "type": "object",
        "additionalProperties": {},
        "description": "A record's data: the fields the kind's schema declares, and anything else a caller stored."
      },
      "ValopayRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "merchantId": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "amountKobo": {
            "type": "integer"
          },
          "customerId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/RecordData"
          }
        },
        "required": [
          "id",
          "merchantId",
          "kind",
          "name",
          "status",
          "reference",
          "amountKobo",
          "customerId",
          "createdAt",
          "updatedAt",
          "data"
        ],
        "description": "A stored record of any kind, with its lender, status, reference, amount in kobo and data."
      },
...
      "Merchant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "shortName": {
            "type": "string"
          },
          "segment": {
            "type": "string"
          },
          "mode": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "monthlyVolume": {
            "type": "integer"
          },
          "killSwitch": {
            "type": "boolean"
          },
          "preDataReady": {
            "type": "boolean"
          },
          "preLiveReady": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "shortName",
          "segment",
          "mode",
          "status",
          "provider",
          "monthlyVolume",
          "killSwitch",
          "preDataReady",
          "preLiveReady"
        ],
        "description": "A lender: its mode (observation or instruction), provider, volume, kill switch and readiness flags."
      },
...
      "ActionInput": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string"
          },
          "recordId": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/RecordData"
          },
          "expectedUpdatedAt": {
            "type": "string"
          }
        },
        "required": [
          "action"
        ],
        "description": "An action to run: its name, the record it applies to, the reason for it and any data it needs."
      },
      "ActionResult": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "record": {
            "$ref": "#/components/schemas/ValopayRecord"
          },
          "data": {
            "$ref": "#/components/schemas/RecordData"
          }
        },
        "required": [
          "message",
          "data"
        ],
        "description": "What an action did, in words, with the record it produced or changed and any data it returns."
      },
      "ImportInput": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string"
          },
          "csv": {
            "type": "string"
          },
          "syntheticOnly": {
            "type": "boolean"
          },
          "commit": {
            "type": "boolean"
          },
          "mapping": {
            "$ref": "#/components/schemas/RecordData"
          },
          "amountUnit": {
            "type": "string",
            "enum": [
              "naira",
              "kobo"
            ],
            "description": "Unit used by source amount values; defaults to kobo for existing API clients. The console requires an explicit choice."
          }
        },
        "required": [
          "kind",
          "csv",
          "syntheticOnly",
          "commit"
        ],
        "description": "A synthetic CSV to preview or commit for one kind, with an optional column mapping."
      },
      "ImportRow": {
        "type": "object",
        "properties": {
          "row": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "row",
          "status",
          "message"
        ],
        "description": "The outcome of one imported row."
      },
      "ImportResult": {
        "type": "object",
        "properties": {
          "valid": {
            "type": "integer"
          },
          "invalid": {
            "type": "integer"
          },
          "imported": {
            "type": "integer"
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImportRow"
            }
          },
          "columns": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "preview": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "row": {
                  "type": "integer"
                },
                "values": {
                  "$ref": "#/components/schemas/RecordData"
                },
                "amountKobo": {
                  "type": "integer"
                }
              },
              "required": [
                "row",
                "values"
              ]
            }
          },
          "skipped": {
            "type": "integer"
          }
        },
        "required": [
          "valid",
          "invalid",
          "imported",
          "rows"
        ],
        "description": "How many rows were valid, invalid and imported, and each row's outcome."
      },
...
      "Settings": {
        "type": "object",
        "properties": {
          "merchant": {
            "$ref": "#/components/schemas/Merchant"
          },
          "settings": {
            "$ref": "#/components/schemas/RecordData"
          },
          "permissions": {
            "$ref": "#/components/schemas/RecordData"
          },
          "integrations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValopayRecord"
            }
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValopayRecord"
            }
          },
          "calendar": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValopayRecord"
            }
          },
          "closeSchedule": {
            "$ref": "#/components/schemas/EffectiveCloseSchedule"
          },
          "revision": {
            "type": "string"
          }
        },
        "required": [
          "merchant",
          "settings",
          "permissions",
          "integrations",
          "members",
          "calendar"
        ],
        "description": "A lender's settings and the caller's permissions, with integrations, members and the business calendar."
      },
      "SettingsInput": {
        "type": "object",
        "properties": {
          "executionStart": {
            "type": "integer"
          },
          "executionEnd": {
            "type": "integer"
          },
          "authorisationMode": {
            "type": "string"
          },
          "contactRoute": {
            "type": "string"
          },
          "minimumTicketKobo": {
            "type": "integer"
          },
          "defaultOwner": {
            "type": "string"
          },
          "policyChangeRequiresConsent": {
            "type": "boolean"
          },
          "unallocatedAlertThreshold": {
            "type": "integer"
          },
          "notificationCostAlertKobo": {
            "type": "integer"
          },
          "closeTime": {
            "type": "string"
          },
          "scheduledCloseEnabled": {
            "type": "boolean"
          },
          "expectedRevision": {
            "type": "string"
          }
        },
        "required": [],
        "description": "The execution settings to change; every field is optional."
      },
      "ExportInput": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "format": {
            "type": "string",
            "enum": [
              "json",
              "csv",
              "pdf"
            ]
          }
        },
        "required": [
          "kind",
          "format"
        ],
        "description": "What to export (a record kind, gate-pack, billing, dispute-pack or customer-pack with a customerId) and in which format."
      },
      "ExportResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "downloadUrl": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "ready",
              "failed"
            ]
          },
          "kind": {
            "type": "string"
          },
          "format": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "requestedAt": {
            "type": "string"
          },
          "attempts": {
            "type": "integer"
          },
          "checksum": {
            "type": "string"
          },
          "generatedAt": {
            "type": "string"
          },
          "byteLength": {
            "type": "integer"
          },
          "generationMs": {
            "type": "integer"
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "downloadUrl"
        ],
        "description": "Saved export job identity, status and retry details. Checksum, generatedAt and file size appear only when ready; the download route rejects unfinished jobs. Optional status retains compatibility with older immediate-export responses."
      },
      "EffectiveCloseSchedule": {
        "type": "object",
        "properties": {
          "time": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "automatic": {
            "type": "boolean"
          },
          "nextAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "runtimeState": {
            "type": "string",
            "enum": [
              "not_started",
              "running",
              "off",
              "stopped"
            ]
          },
          "serviceIssue": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "starting",
              "delayed",
              "failed",
              null
            ]
          },
          "missed": {
            "type": "boolean"
          },
          "overdueMinutes": {
            "type": "integer"
          },
          "lateAfterMinutes": {
            "type": "integer"
          },
          "lastAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "lastTrigger": {
            "type": [
              "string",
              "null"
            ]
          },
          "lastCheckedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "lastErrorAt": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "time",
          "enabled",
          "automatic",
          "nextAt",
          "runtimeState",
          "serviceIssue",
          "missed",
          "overdueMinutes",
          "lateAfterMinutes",
          "lastAt",
          "lastTrigger",
          "lastCheckedAt",
          "lastErrorAt"
        ],
        "description": "Lender schedule combined with the actual scheduler service status. nextAt is present only when automatic closes are available; run history belongs only to this lender."
      },
...
      "ConnectedWorkspace": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "synthetic"
            ]
          },
          "revision": {
            "type": "string"
          },
          "asOf": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "entity": {
            "$ref": "#/components/schemas/RecordData"
          },
          "customers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecordData"
            }
          },
          "consents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValopayRecord"
            }
          },
          "purposes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecordData"
            }
          },
          "gates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecordData"
            }
          },
          "payments": {
            "$ref": "#/components/schemas/RecordData"
          },
          "credit": {
            "$ref": "#/components/schemas/RecordData"
          },
          "cash": {
            "$ref": "#/components/schemas/RecordData"
          }
        },
        "required": [
          "mode",
          "revision",
          "asOf",
          "role",
          "entity",
          "customers",
          "consents",
          "purposes",
          "gates",
          "payments",
          "credit",
          "cash"
        ],
        "description": "Synthetic connected workspace. Credit and Cash views carry evidence, permissions and refusal states. Every gate has liveEnabled false. No read creates sample records."
      },
      "ConnectedActionInput": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "maxLength": 80
          },
          "recordId": {
            "type": "string",
            "maxLength": 100
          },
          "reason": {
            "type": "string",
            "minLength": 8,
            "maxLength": 500
          },
          "data": {
            "$ref": "#/components/schemas/RecordData"
          },
          "expectedRevision": {
            "type": "string",
            "maxLength": 80
          }
        },
        "required": [
          "action",
          "reason",
          "expectedRevision"
        ],
        "additionalProperties": false,
        "description": "Action-specific data is validated by the server. Names use consent, payment, credit or cash prefixes. Every action requires a current whole-workspace revision and a reason. No input can enable live routes."
      },
      "ConnectedActionResult": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "record": {
            "$ref": "#/components/schemas/RecordData"
          },
          "mode": {
            "type": "string",
            "enum": [
              "synthetic"
            ]
          },
          "externalInstructionPerformed": {
            "type": "boolean",
            "const": false
          }
        },
        "required": [
          "message",
          "record",
          "mode",
          "externalInstructionPerformed"
        ],
        "description": "Committed sample operation. Cash actions include their record and outcome inside record; use the refreshed workspace view for display. A receipt is evidence from the server simulator only."
      }
    }
  }
}